+++ src/lib/components/forms/SelectInput.tsx
@@ -6,22 +6,33 @@ import {
   Popover,
   PopoverTrigger,
   PopoverContent,
-  Box,
   useDisclosure,
   useOutsideClick,
+  Flex,
+  InputLeftElement,
 } from "@chakra-ui/react";
 import type { MutableRefObject, ReactNode } from "react";
-import { useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
+import type { IconType } from "react-icons/lib";
 import { MdArrowDropDown } from "react-icons/md";
 
-const ITEM_HEIGHT = 57;
+import type { Option } from "lib/types";
 
-interface SelectInputProps {
+const ITEM_HEIGHT = 56;
+
+interface SelectInputProps<T extends string> {
   formLabel?: string;
-  options: { label: string; value: string; disabled: boolean }[];
-  onChange: (newVal: string) => void;
+  options: {
+    label: string;
+    value: T;
+    disabled: boolean;
+    icon?: IconType;
+    iconColor?: string;
+  }[];
+  onChange: (newVal: T) => void;
   placeholder?: string;
   initialSelected: string;
+  hasDivider?: boolean;
 }
 
 interface SelectItemProps {
@@ -32,40 +43,49 @@ interface SelectItemProps {
 
 const SelectItem = ({ children, onSelect, disabled }: SelectItemProps) => {
   return (
-    <Box
-      p={4}
+    <Flex
+      px={4}
+      py={2}
       onClick={onSelect}
       color="text.main"
       transition="all .2s"
       cursor="pointer"
+      gap={2}
       aria-disabled={disabled}
       _hover={{ bg: "gray.800" }}
       _disabled={{ opacity: 0.4, pointerEvents: "none" }}
     >
       {children}
-    </Box>
+    </Flex>
   );
 };
 
-export const SelectInput = ({
+export const SelectInput = <T extends string>({
   formLabel,
   options,
   onChange,
   placeholder = "",
   initialSelected,
-}: SelectInputProps) => {
+  hasDivider = false,
+}: SelectInputProps<T>) => {
   const optionRef = useRef() as MutableRefObject<HTMLElement>;
   const { isOpen, onClose, onOpen } = useDisclosure();
-
+  const inputRef = useRef() as MutableRefObject<HTMLInputElement>;
   const [selected, setSelected] = useState(
-    () => options.find((asset) => asset.value === initialSelected)?.label ?? ""
+    () => options.find((item) => item.value === initialSelected)?.label ?? ""
   );
-
+  const [inputRefWidth, setInputRefWidth] = useState<Option<number>>();
   useOutsideClick({
     ref: optionRef,
     handler: () => isOpen && onClose(),
   });
+  const selectedOption = options.find((item) => item.label === selected);
 
+  useEffect(() => {
+    if (inputRef.current) {
+      setInputRefWidth(inputRef.current.clientWidth);
+    }
+  }, [inputRef]);
   return (
     <Popover placement="bottom-start" isOpen={isOpen}>
       <PopoverTrigger>
@@ -92,6 +112,15 @@ export const SelectInput = ({
           }}
         >
           <div className="form-label">{formLabel}</div>
+          {selectedOption?.icon && (
+            <InputLeftElement pointerEvents="none" h="full">
+              <Icon
+                as={selectedOption.icon}
+                color={selectedOption.iconColor}
+                fontSize="20px"
+              />
+            </InputLeftElement>
+          )}
           <Input
             size="lg"
             textAlign="start"
@@ -99,6 +128,8 @@ export const SelectInput = ({
             value={selected || placeholder}
             fontSize="14px"
             color={selected ? "text.main" : "text.dark"}
+            ref={inputRef}
+            pl={selectedOption?.icon ? 9 : 4}
           />
           <InputRightElement pointerEvents="none" h="full">
             <Icon as={MdArrowDropDown} color="text.dark" fontSize="24px" />
@@ -109,7 +140,7 @@ export const SelectInput = ({
         ref={optionRef}
         border="unset"
         bg="gray.900"
-        w="200px"
+        w={inputRefWidth}
         maxH={`${ITEM_HEIGHT * 4}px`}
         overflow="scroll"
         borderRadius="4px"
@@ -118,12 +149,12 @@ export const SelectInput = ({
         }}
         sx={{
           "> div:not(:last-of-type)": {
-            borderBottom: "1px solid",
-            borderBottomColor: "divider.main",
+            borderBottom: hasDivider && "1px solid",
+            borderBottomColor: hasDivider && "divider.main",
           },
         }}
       >
-        {options.map(({ label, value, disabled }) => (
+        {options.map(({ label, value, disabled, icon, iconColor }) => (
           <SelectItem
             key={value}
             onSelect={() => {
@@ -133,6 +164,7 @@ export const SelectInput = ({
             }}
             disabled={disabled}
           >
+            {icon && <Icon as={icon} boxSize={5} color={iconColor} />}
             {label}
           </SelectItem>
         ))}